app/requirements.txt
annotated-types==0.7.0 ; python_version >= "3.11"
anyio==4.9.0 ; python_version >= "3.11"
attrs==25.3.0 ; python_version >= "3.11"
browsermob-proxy==0.8.0 ; python_version >= "3.11"
certifi==2025.8.3 ; python_version >= "3.11"
cffi==1.17.1 ; os_name == "nt" and implementation_name != "pypy" and python_version >= "3.11"
charset-normalizer==3.4.2 ; python_version >= "3.11"
click==8.2.1 ; python_version >= "3.11"
colorama==0.4.6 ; python_version >= "3.11" and platform_system == "Windows"
fastapi==0.116.1 ; python_version >= "3.11"
h11==0.16.0 ; python_version >= "3.11"
idna==3.10 ; python_version >= "3.11"
outcome==1.3.0.post0 ; python_version >= "3.11"
packaging==25.0 ; python_version >= "3.11"
pycparser==2.22 ; os_name == "nt" and implementation_name != "pypy" and python_version >= "3.11"
pydantic-core==2.33.2 ; python_version >= "3.11"
pydantic==2.11.7 ; python_version >= "3.11"
pysocks==1.7.1 ; python_version >= "3.11"
python-dotenv==1.1.1 ; python_version >= "3.11"
requests==2.32.4 ; python_version >= "3.11"
selenium==4.34.2 ; python_version >= "3.11"
sniffio==1.3.1 ; python_version >= "3.11"
sortedcontainers==2.4.0 ; python_version >= "3.11"
starlette==0.47.2 ; python_version >= "3.11"
trio-websocket==0.12.2 ; python_version >= "3.11"
trio==0.30.0 ; python_version >= "3.11"
typing-extensions==4.14.1 ; python_version >= "3.11"
typing-inspection==0.4.1 ; python_version >= "3.11"
undetected-chromedriver==3.5.5 ; python_version >= "3.11"
urllib3==2.5.0 ; python_version >= "3.11"
uvicorn==0.35.0 ; python_version >= "3.11"
webdriver-manager==4.0.2 ; python_version >= "3.11"
websocket-client==1.8.0 ; python_version >= "3.11"
websockets==15.0.1 ; python_version >= "3.11"
wsproto==1.2.0 ; python_version >= "3.11"
